+++ POSCoreTests/CalendarPlanLogic/CalendarWorkCreatorTests/CreateCalendarWorkShould.cs
@@ -52,7 +52,7 @@ namespace POSCoreTests.CalendarPlanLogic.CalendarWorkCreatorTests
         [Test]
         public void ReturnCalendarWork_InWhichSetTotalCostIncludingContructionAndInstallationWorks_ToTotalCostMinusEquipmentCostAndOtherProductsCost()
         {
-            var estimateWork = new EstimateWork("ЭЛЕКТРОХИМИЧЕСКАЯ ЗАЩИТА", 2.027, 1.541, 55.464, 0);
+            var estimateWork = new EstimateWork("ЭЛЕКТРОХИМИЧЕСКАЯ ЗАЩИТА", (decimal)2.027, (decimal)1.541, (decimal)55.464, 0);
 
             var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new int[] { 0 });
 
@@ -72,7 +72,7 @@ namespace POSCoreTests.CalendarPlanLogic.CalendarWorkCreatorTests
         [Test]
         public void CreateCalendarPeriod()
         {
-            var estimateWork = new EstimateWork("ЭЛЕКТРОХИМИЧЕСКАЯ ЗАЩИТА", 0, 0, 11.324, 0);
+            var estimateWork = new EstimateWork("ЭЛЕКТРОХИМИЧЕСКАЯ ЗАЩИТА", 0, 0, (decimal)11.324, 0);
             var initialDate = new DateTime(1999, 9, 21);
             var percentages = new int[] { 1 };
